+++ NEWS.md
@@ -1,5 +1,7 @@
 # fuj (development version)
 
+* adds `verbose()` for conditional message printing [#29](https://github.com/jmbarbone/fuj/issues/29)
+
 # fuj 0.1.3
 
 * `list0()` now accepts `named` arguments in `...` [#25](https://github.com/jmbarbone/fuj/issues/25)